@GoldFireGames - podstawowa walidacja w userDTO, - obsługa wyjątków dla walidacji w ApplicationExceptionHandler - dopisanie obsługi wyjątku w UserService dla @DeleteMapping - poprawienie pól usera (dodanie firstName)

+++ currencyhub/src/main/java/com/truejacobg/currencyhub/user/UserService.java
@@ -3,11 +3,13 @@ package com.truejacobg.currencyhub.user;
 import com.truejacobg.currencyhub.user.dto.CreateUserResponseDTO;
 import com.truejacobg.currencyhub.user.dto.UserDTO;
 import com.truejacobg.currencyhub.user.entity.UserEntity;
+import jakarta.validation.constraints.Email;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.Optional;
 // prawdziwy backend metody walidacje, zabawa z bazą, prawdziwe mięso, tutaj bedzie cały kod
 // musi implementować repo, 90% będzie tutaj, autoryzacje, zabezpieczenia, uruchomienia, rzucanie błędów
 //
@@ -18,12 +20,11 @@ public class UserService {
     private UserRepository userRepository;
 
     public CreateUserResponseDTO createUser(UserDTO userDTO) {
-        UserEntity userEntity = new UserEntity(null, userDTO.getName(), userDTO.getAuthCode(), userDTO.getSurname(), userDTO.getEmail(), LocalDateTime.now());
+        UserEntity userEntity = new UserEntity(null,userDTO.getFirstName(), userDTO.getName(), userDTO.getAuthCode(), userDTO.getSurname(), userDTO.getEmail(), LocalDateTime.now());
         userRepository.save(userEntity);
 
         return new CreateUserResponseDTO("ok", HttpStatus.ACCEPTED);
     }
-
 
     public CreateUserResponseDTO getUser(String userEmail) {
         UserEntity user = userRepository.findByEmail(userEmail);
@@ -41,6 +42,7 @@ public class UserService {
         if (user == null) {
             return new CreateUserResponseDTO("User has not been in db", HttpStatus.NOT_FOUND);
         } else {
+            user.setFirstName(userDTO.getFirstName());
             user.setName(userDTO.getName());
             user.setSurname(userDTO.getSurname());
             user.setEmail(userDTO.getEmail());
@@ -52,8 +54,14 @@ public class UserService {
         }
     }
 
-    public CreateUserResponseDTO deleteUser(String email) {
-        userRepository.deleteByEmail(email); // TODO: dodać wyjątek
-        return new CreateUserResponseDTO("User has been deleted", HttpStatus.OK);
+      public CreateUserResponseDTO deleteUser(String email) {
+        Optional<UserEntity> user = Optional.ofNullable(userRepository.findByEmail(email));
+        if(!user.isPresent()){
+            return new CreateUserResponseDTO("User has not been in db", HttpStatus.NOT_FOUND);
+        }else {
+            userRepository.deleteByEmail(email);
+            return new CreateUserResponseDTO("User has been deleted", HttpStatus.OK);
+        }
     }
+
 }
